tests/wpt/tests/html/semantics/forms/the-select-element/select-pseudo-light-dismiss-invalidation.tentative.html

<!DOCTYPE html>
<link rel=author href="mailto:jarhar@chromium.org">
<link rel=help href="http://crbug.com/1429839">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>

<select id=select>
  <option id=optone>one</option>
  <option id=opttwo>two</option>
</select>
<style>
select {
  background-color: rgb(0, 0, 255);
}
select:closed {
  background-color: rgb(0, 255, 0);
}
select:open {
  background-color: rgb(255, 0, 0);
}
select, ::picker(select) {
  appearance: base-select;
}
</style>
<button id=button>option</button>

<script>
promise_test(async () => {
  assert_equals(getComputedStyle(select).backgroundColor, 'rgb(0, 255, 0)',
    'The select should match :closed at the start of the test.');

  await test_driver.click(select);
  assert_equals(getComputedStyle(select).backgroundColor, 'rgb(255, 0, 0)',
    'The select should match :open when opened.');

  await test_driver.click(opttwo);
  assert_equals(getComputedStyle(select).backgroundColor, 'rgb(0, 255, 0)',
    'The select should match :closed after clicking an option.');

  await test_driver.click(select);
  assert_equals(getComputedStyle(select).backgroundColor, 'rgb(255, 0, 0)',
    'The select should match :open when reopened.');

  await test_driver.click(button);
  assert_equals(getComputedStyle(select).backgroundColor, 'rgb(0, 255, 0)',
    'The select should match :closed after light dismiss.');
}, 'select should not match :open when light dismissed.');
</script>
